Refine the struct of the code
